+++ public/index.html
@@ -1,27 +1,31 @@
+
 <!doctype html>
 <html lang="en">
   <head>
-    <title>CS4241 Assignment 2</title>
+    <title>CS4241 Assignment 3</title>
     <meta charset="utf-8">
     <script src="js/main.js"></script>
-    <link rel="stylesheet" href="css/main.css">
+    <!-- <link rel="stylesheet" href="css/main.css"> -->
+    <link rel="stylesheet" href="css/bootstrap.min.css">
     
     <!-- get rid of favicon error -->
     <link rel="icon" href="data:;base64,iVBORw0KGgo=">
   </head>
   <body>
-    <p class="p_class">CAR MODEL FORM</p>
+    <p class="p_class">USER DATA FORM</p>
+
+    <div class="each_data">
+        <div> User Name</div>
+        <div id="user_name_field"> </div>
+    </div>
+
     <form id="login_form">
-      <br/><label>Car Model:</label><br>
-      <input type="text" id="model" value="Default">
-      <br/><label>Year of Production:</label><br>
-      <input type="number" id="year" value=2000>
-      <br/><label>MPG:</label><br/>
-      <input type="number" id="mpg" value=10>
-      <br/><label>Shelf Life:</label><br/>
-      <input type="number" id="shelf_life" value=10>
+      <br/><label>Ready:</label><br>
+      <input type="checkbox" id="ready" value="True">
+      <br/><label>Age:</label><br>
+      <input type="number" id="age" value=0>
       <br/>
-      <button>submit</button>
+      <button id="add_to_db_button">Add</button>
     </form>
     <br/>
 <!--
