--- public/index_2.html
@@ -1,42 +1,0 @@
-
-<!doctype html>
-<html lang="en">
-  <head>
-    <title>CS4241 Assignment 3</title>
-    <meta charset="utf-8">
-    <script src="js/main_2.js"></script>
-    <!-- <link rel="stylesheet" href="css/main.css"> -->
-    <link rel="stylesheet" href="css/bootstrap.min.css">
-    
-    <!-- get rid of favicon error -->
-    <link rel="icon" href="data:;base64,iVBORw0KGgo=">
-  </head>
-  <body>
-    <p class="p_class">USER DATA FORM</p>
-
-    <div class="each_data">
-        <div> User Name</div>
-        <div id="user_name_field"> </div>
-    </div>
-
-    <form id="login_form">
-      <br/><label>Ready:</label><br>
-      <input type="checkbox" id="ready" value="True">
-      <br/><label>Age:</label><br>
-      <input type="number" id="age" value=0>
-      <br/>
-      <button id="add_to_db_button">Add</button>
-    </form>
-    <br/>
-<!-- 
-    <div id="show_data">
-      <div class="each_data">
-        <div> Car Model</div>
-        <div> Year of Production</div>
-        <div> MPG</div>
-        <div> Shelf Life</div>
-        <div> Year of Expirationi</div>
-      </div>
-    </div> -->
-  </body>
-</html>
