--- public/login_2.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Login Page</title>
-    <!-- <link rel="stylesheet" type="text/css" href="css/login.css"> -->
-    <link rel="stylesheet" type="text/css" href="css/bootstrap.min.css">
-</head>
-<body>
-    <div class="login-container">
-        <h2>Login Form</h2>
-        <input type="text" id="username" placeholder="Username">
-        <input type="password" id="password" placeholder="Password">
-        <button id="login_button">Login</button>
-    </div>
-    <script src="js/login_3.js"></script>
-</body>
-</html>
